+++ resolwe_bio/processes/reads_processing/cutadapt_corall.py
@@ -25,7 +25,7 @@ class CutadaptCorallSingle(Process):
     slug = "cutadapt-corall-single"
     name = "Cutadapt (Corall RNA-Seq, single-end)"
     process_type = "data:reads:fastq:single:cutadapt:"
-    version = "1.4.1"
+    version = "1.4.2"
     category = "FASTQ processing"
     scheduling_class = SchedulingClass.BATCH
     entity = {"type": "sample"}
@@ -35,7 +35,7 @@ class CutadaptCorallSingle(Process):
             "docker": {"image": "public.ecr.aws/genialis/resolwebio/rnaseq:6.0.0"},
         },
         "resources": {
-            "cores": 10,
+            "cores": 4,
             "memory": 16384,
         },
     }
@@ -198,7 +198,7 @@ class CutadaptCorallPaired(Process):
     slug = "cutadapt-corall-paired"
     name = "Cutadapt (Corall RNA-Seq, paired-end)"
     process_type = "data:reads:fastq:paired:cutadapt:"
-    version = "1.3.1"
+    version = "1.3.2"
     category = "FASTQ processing"
     scheduling_class = SchedulingClass.BATCH
     entity = {"type": "sample"}
@@ -208,7 +208,7 @@ class CutadaptCorallPaired(Process):
             "docker": {"image": "public.ecr.aws/genialis/resolwebio/rnaseq:6.0.0"},
         },
         "resources": {
-            "cores": 10,
+            "cores": 4,
             "memory": 16384,
         },
     }